python/requirements.txt
fastapi==0.103.2
uvicorn==0.27.1
sqlalchemy>=1.4.24,<1.5
clickhouse-driver==0.2.6
clickhouse-sqlalchemy==0.2.5
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
redis==5.0.1
alembic==1.13.1
email-validator==2.1.0 